Moving all popup creation functionality into the popup trait module
