+++ frontend/src/models/chicago.rs
@@ -4,12 +4,6 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use wasm_bindgen::JsCast;
-use web_sys::{
-    Document, Element, HtmlElement, HtmlTableCellElement, HtmlTableElement, HtmlTableRowElement,
-};
-
-use crate::errors::StaccError;
 
 /// Holds the Chicago ShotSpotter Alert and Victims of Homicide and Non-Fatal Shootings data.
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
@@ -131,99 +125,6 @@ pub struct ShotData {
     pub zip_code: String,
 }
 
-impl ShotData {
-    /// Create the header for the popup.
-    pub fn create_popup_header(
-        &self,
-        document: &Document,
-        popup_content: &Element,
-    ) -> Result<(), StaccError> {
-        let header = document.create_element("h5")?;
-        header.set_inner_html(&self.incident_type_description);
-        popup_content.append_child(&header)?;
-
-        Ok(())
-    }
-
-    /// Create the metadata subtitle/secondary header for the popup.
-    pub fn create_metadata_subtitle(
-        &self,
-        document: &Document,
-        popup_content: &Element,
-        timestamp: String,
-    ) -> Result<(), StaccError> {
-        let meta_subtitle = document.create_element("small")?;
-        meta_subtitle.set_inner_html(&format!(
-            "{} | {} {}",
-            timestamp, self.community_area, self.zip_code
-        ));
-        popup_content.append_child(&meta_subtitle)?;
-
-        Ok(())
-    }
-
-    /// Create the info table containing the block on which the shots were detected, as well as the
-    /// number of rounds detected if more than one gunshot was detected.
-    pub fn create_info_table(
-        &self,
-        document: &Document,
-        popup_content: &Element,
-    ) -> Result<(), StaccError> {
-        let info_table = document
-            .create_element("table")?
-            .dyn_into::<HtmlTableElement>()?;
-        info_table.set_class_name("marker-popup-table");
-
-        let block_row = document
-            .create_element("tr")?
-            .dyn_into::<HtmlTableRowElement>()?;
-
-        let block_title = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        block_title.set_class_name("marker-popup-table-cell");
-        block_title.set_inner_html("<b>Block</b>");
-
-        let block_value = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        block_value.set_class_name("marker-popup-table-cell");
-        block_value.set_inner_html(&self.block.trim_end_matches(','));
-
-        block_row.append_child(&block_title)?;
-        block_row.append_child(&block_value)?;
-
-        info_table.append_child(&block_row)?;
-
-        if self.incident_type_description.to_lowercase() == "multiple gunshots" {
-            let rounds_row = document
-                .create_element("tr")?
-                .dyn_into::<HtmlTableRowElement>()?;
-
-            let rounds_title = document
-                .create_element("td")?
-                .dyn_into::<HtmlTableCellElement>()?;
-            rounds_title.set_class_name("marker-popup-table-cell");
-            rounds_title.set_inner_html("<b>Rounds</b>");
-
-            let rounds_value = document
-                .create_element("td")?
-                .dyn_into::<HtmlTableCellElement>()?;
-            rounds_value.set_class_name("marker-popup-table-cell");
-            rounds_value.set_inner_html(&self.rounds);
-
-            rounds_row.append_child(&rounds_title)?;
-            rounds_row.append_child(&rounds_value)?;
-
-            info_table.append_child(&rounds_row)?;
-        }
-
-        popup_content.append_child(&info_table.dyn_into::<HtmlElement>()?.into())?;
-
-        Ok(())
-    }
-}
-
 /// Holds the most interesting data from the violence data.
 /// Information for these fields reference the Socrata API documentation here:
 /// [Socrata Violence API Documentation](https://dev.socrata.com/foundry/data.cityofchicago.org/gumc-mgzr)
@@ -262,195 +163,6 @@ pub struct ViolenceData {
 }
 
 impl ViolenceData {
-    /// Create the header for the popup.
-    pub fn create_popup_header(
-        &self,
-        crime_description: String,
-        document: &Document,
-        popup_content: &Element,
-    ) -> Result<(), StaccError> {
-        let header = document.create_element("h5")?;
-        header.set_inner_html(&crime_description);
-        popup_content.append_child(&header)?;
-
-        Ok(())
-    }
-
-    /// Create the metadata subtitle/secondary header for the popup.
-    pub fn create_metadata_subtitle(
-        &self,
-        document: &Document,
-        popup_content: &Element,
-        timestamp: String,
-    ) -> Result<(), StaccError> {
-        let meta_subtitle = document.create_element("small")?;
-        meta_subtitle.set_inner_html(&format!(
-            "{} | {} {}",
-            timestamp, self.community_area, self.zip_code
-        ));
-        popup_content.append_child(&meta_subtitle)?;
-
-        Ok(())
-    }
-
-    /// Create the info table containing the block on which the shots were detected, as well as the
-    /// number of rounds detected if more than one gunshot was detected.
-    pub fn create_info_table(
-        &self,
-        document: &Document,
-        popup_content: &Element,
-    ) -> Result<(), StaccError> {
-        let info_table = document
-            .create_element("table")?
-            .dyn_into::<HtmlTableElement>()?;
-        info_table.set_class_name("marker-popup-table");
-
-        let gunshot_injury_row = self.build_gunshot_injury_row(document)?;
-        info_table.append_child(&gunshot_injury_row)?;
-
-        let location_description_row = self.build_location_description_row(document)?;
-        info_table.append_child(&location_description_row)?;
-
-        let race_row = self.build_race_row(document)?;
-        info_table.append_child(&race_row)?;
-
-        let age_row = self.build_age_row(document)?;
-        info_table.append_child(&age_row)?;
-
-        let sex_row = self.build_sex_row(document)?;
-        info_table.append_child(&sex_row)?;
-
-        popup_content.append_child(&info_table.dyn_into::<HtmlElement>()?.into())?;
-
-        Ok(())
-    }
-
-    /// Build the gunshot injury row.
-    fn build_gunshot_injury_row(
-        &self,
-        document: &Document,
-    ) -> Result<HtmlTableRowElement, StaccError> {
-        let gunshot_injury_row = document
-            .create_element("tr")?
-            .dyn_into::<HtmlTableRowElement>()?;
-
-        let gunshot_injury_title = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        gunshot_injury_title.set_class_name("marker-popup-table-cell");
-        gunshot_injury_title.set_inner_html("<b>Gunshot injury?</b>");
-
-        let gunshot_injury_value = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        gunshot_injury_value.set_class_name("marker-popup-table-cell");
-        gunshot_injury_value.set_inner_html(&self.gunshot_injury_i);
-
-        gunshot_injury_row.append_child(&gunshot_injury_title)?;
-        gunshot_injury_row.append_child(&gunshot_injury_value)?;
-
-        Ok(gunshot_injury_row)
-    }
-
-    /// Build the location description row.
-    fn build_location_description_row(
-        &self,
-        document: &Document,
-    ) -> Result<HtmlTableRowElement, StaccError> {
-        let location_description_row = document
-            .create_element("tr")?
-            .dyn_into::<HtmlTableRowElement>()?;
-
-        let location_title = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        location_title.set_class_name("marker-popup-table-cell");
-        location_title.set_inner_html("<b>Location</b>");
-
-        let location_value = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        location_value.set_class_name("marker-popup-table-cell");
-        location_value.set_inner_html(&self.location_description);
-
-        location_description_row.append_child(&location_title)?;
-        location_description_row.append_child(&location_value)?;
-
-        Ok(location_description_row)
-    }
-
-    /// Build the victim race row.
-    fn build_race_row(&self, document: &Document) -> Result<HtmlTableRowElement, StaccError> {
-        let race_row = document
-            .create_element("tr")?
-            .dyn_into::<HtmlTableRowElement>()?;
-
-        let race_title = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        race_title.set_class_name("marker-popup-table-cell");
-        race_title.set_inner_html("<b>Victim race</b>");
-
-        let race_value = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        race_value.set_class_name("marker-popup-table-cell");
-        race_value.set_inner_html(&self.race);
-
-        race_row.append_child(&race_title)?;
-        race_row.append_child(&race_value)?;
-
-        Ok(race_row)
-    }
-
-    /// Build the victim age row.
-    fn build_age_row(&self, document: &Document) -> Result<HtmlTableRowElement, StaccError> {
-        let age_row = document
-            .create_element("tr")?
-            .dyn_into::<HtmlTableRowElement>()?;
-
-        let age_title = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        age_title.set_class_name("marker-popup-table-cell");
-        age_title.set_inner_html("<b>Victim age</b>");
-
-        let age_value = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        age_value.set_class_name("marker-popup-table-cell");
-        age_value.set_inner_html(&self.age);
-
-        age_row.append_child(&age_title)?;
-        age_row.append_child(&age_value)?;
-
-        Ok(age_row)
-    }
-
-    /// Build the victim sex row.
-    fn build_sex_row(&self, document: &Document) -> Result<HtmlTableRowElement, StaccError> {
-        let sex_row = document
-            .create_element("tr")?
-            .dyn_into::<HtmlTableRowElement>()?;
-
-        let sex_title = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        sex_title.set_class_name("marker-popup-table-cell");
-        sex_title.set_inner_html("<b>Victim sex</b>");
-
-        let sex_value = document
-            .create_element("td")?
-            .dyn_into::<HtmlTableCellElement>()?;
-        sex_value.set_class_name("marker-popup-table-cell");
-        sex_value.set_inner_html(&self.sex);
-
-        sex_row.append_child(&sex_title)?;
-        sex_row.append_child(&sex_value)?;
-
-        Ok(sex_row)
-    }
-
     /// Get the crime description that matches the `self.incident_iucr_cd` code.
     /// Codes were found at this [Chicago Police Crime Details page](https://gis.chicagopolice.org/pages/crime_details).
     ///
